requirements.txt
asgiref==3.7.2
Django==4.2.5
django-restframework==0.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
install==1.3.5
packaging==23.2
Pillow==10.0.1
psycopg==3.1.12
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing-extensions==4.7.1
uritemplate==4.1.1
